+++ tests/fixtures/config/wrong_branches.py
@@ -15,3 +15,16 @@ def check_branches_in_function():
             5
     elif 6:  # error here
         print()
+
+
+async def async_check_branches_in_function():
+    if 1:
+        if 2:
+            print()
+    elif 3:
+        try:
+            4
+        except:
+            5
+    elif 6:  # error here
+        print()