+++ tests/fixtures/config/wrong_expressions.py
@@ -16,3 +16,17 @@ def check_expressions_in_function():
     x
 
     1 / x  # error here
+
+
+async def async_check_expressions_in_function():
+    1 / 0
+    print(1)
+    x
+    1 / 0
+    print(1)
+    x
+    1 / 0
+    print(1)
+    x
+
+    1 / x  # error here